Extract functions from properties panel
