+++ docs/refactor-architecture-plan.md
@@ -250,7 +250,7 @@ Checklist:
 Current sprint focus:
 
 - [x] Refactor `GcodeOptionsDialog.tsx`
-- [ ] Refactor `PropertiesPanel.tsx`
+- [ ] Refactor `PropertiesPanel.tsx` (in progress: toolpath metrics helpers extracted)
 - [ ] Refactor `Toolbar.tsx`
 - [ ] Start `canvasStore.ts` slice extraction
 
@@ -259,6 +259,7 @@ Done this sprint:
 - [x] Initialize architecture refactor plan and checklist
 - [x] Start `GcodeOptionsDialog.tsx` refactor (prefs + validators extracted, behavior verified)
 - [x] Extract `OptionsSection`, `OutputSection`, and `CustomGcodeSection`
+- [x] Extract PropertiesPanel toolpath metrics helpers (`formatBytes`, `formatDuration`, `estimateDuration`)
 
 ## Progress Log
 
@@ -269,6 +270,7 @@ Use this section to track completed steps with date and PR/commit references.
 - 2026-04-06: Verified `GcodeOptionsDialog` behavior parity with `npm run test -- tests/component/GcodeOptionsDialog.test.tsx` (32/32 tests passing).
 - 2026-04-06: Extracted `PathsSection` from `GcodeOptionsDialog` into `features/gcode-options/components/PathsSection.tsx`; targeted dialog tests remain 32/32 passing.
 - 2026-04-06: Extracted `OptionsSection`, `OutputSection`, and nested `CustomGcodeSection`; targeted dialog tests remain 32/32 passing.
+- 2026-04-06: Extracted `PropertiesPanel` toolpath metric helpers to `features/properties-panel/utils/toolpathMetrics.ts`; targeted panel tests remain 74/74 passing.
 
 ## Update Rule
